fix: build tools run.sh

+++ tools/build/run.sh
@@ -21,7 +21,7 @@ DOCKERHUB_TOKEN=${DOCKERHUB_TOKEN:-}
 # eg1: log "error" "Failed to xxx"
 # eg2: log "xxx complete!"
 function log() {
-  local logLevel="\033[33mINFO\033[0m "
+  local logLevel=" \033[33mINFO\033[0m"
   local logContent=$1
   if [[ $# > 1 ]]; then
     logLevel=$1
@@ -59,6 +59,7 @@ export DOCKERHUB_TOKEN='abc'
 
 Usage: ./$(basename $0) [OPTIONS] [arg1] [arg2] ...
     version                             Print maven project POM version.
+    gpg-verify                          Verifying for GPG.
     build-maven                         Build with Maven.
     build-deploy                        Build and deploy to Maven central.
     build-image                         Build component images.
@@ -92,7 +93,6 @@ function do_build_maven() {
     [ -z "$MAVEN_OPTS" ] && MAVEN_OPTS=$DEFAULT_MAVEN_OPTS
     $BASE_DIR/mvnw \
     $MAVEN_CLI_OPTS \
-    --settings .github/mvn-settings.xml
     -Dmaven.repo.local=$HOME/.m2/repository \
     -Dmaven.test.skip=true \
     -DskipTests \
@@ -103,14 +103,9 @@ function do_build_maven() {
 }
 
 function do_configure_gpg() {
-    log "INFO" "Configuring for GPG keys ..."
+    log "Configuring for GPG keys ..."
 
-    # Check for gpg keys.
-    if [[ -z "$MAVEN_GPG_PRIVATE_KEY" || -z "$MAVEN_GPG_PASSPHRASE" ]]; then
-      logErr "The environment variable MAVEN_GPG_PRIVATE_KEY or MAVEN_GPG_PASSPHRASE is missing."; exit 1
-    fi
-
-    # Check for supported gpg version.
+    # Check for supported GPG version.
     gpg_version=$(gpg --version | head -1 | grep -iEo '(([0-9]+)\.([0-9]+)\.([0-9]+))') # eg: 2.2.19
     gpg_version_major=$(echo $gpg_version | awk -F '.' '{print $1}')
     gpg_version_minor=$(echo $gpg_version | awk -F '.' '{print $2}')
@@ -119,34 +114,44 @@ function do_configure_gpg() {
       logErr "The GPG version must >= $gpg_version_major.$gpg_version_minor.x"; exit 1
     fi
 
-    \rm -rf ~/.gnupg/; mkdir -p ~/.gnupg/private-keys-v1.d/; chmod -R 700 ~/.gnupg/
-    echo -n "$MAVEN_GPG_PRIVATE_KEY" > /tmp/private.key
+    # Check for GPG keys. (Skip if already generated)
+    if [[ ! -f "$HOME/.gnupg/pubring.kbx" ]]; then
+        if [[ -z "$MAVEN_GPG_PRIVATE_KEY" ]]; then
+            logErr "The environment variable MAVEN_GPG_PRIVATE_KEY is missing."; exit 1
+        fi
 
-    #logDebug "----- Print GPG secret key (debug) -----"
-    #cat /tmp/private.key
+        \rm -rf ~/.gnupg/; mkdir -p ~/.gnupg/private-keys-v1.d/; chmod -R 700 ~/.gnupg/
+        echo -n "$MAVEN_GPG_PRIVATE_KEY" > /tmp/private.key
 
-    # FIXED:https://github.com/keybase/keybase-issues/issues/2798#issue-205008630
-    #export GPG_TTY=$(tty) # Notice: github action the VM instance no tty.
+        #logDebug "----- Print GPG secret key (debug) -----"
+        #cat /tmp/private.key
 
-    # FIXED:https://bbs.archlinux.org/viewtopic.php?pid=1691978#p1691978
-    # FIXED:https://github.com/nodejs/docker-node/issues/922
-    # Note that since Version 2.0 this passphrase is only used if the option --batch has also
-    # been given. Since Version 2.1 the --pinentry-mode also needs to be set to loopback.
-    # see:https://www.gnupg.org/documentation/manuals/gnupg/GPG-Esoteric-Options.html#index-allow_002dsecret_002dkey_002dimport
-    gpg2 -v --pinentry-mode loopback --batch --secret-keyring ~/.gnupg/secring.gpg --import /tmp/private.key
+        # FIXED:https://github.com/keybase/keybase-issues/issues/2798#issue-205008630
+        #export GPG_TTY=$(tty) # Notice: github action the VM instance no tty.
 
-    logDebug "Cleanup to /tmp/private.key ..."
-    \rm -rf /tmp/private.key
-    ls -al ~/.gnupg/
+        # FIXED:https://bbs.archlinux.org/viewtopic.php?pid=1691978#p1691978
+        # FIXED:https://github.com/nodejs/docker-node/issues/922
+        # Note that since Version 2.0 this passphrase is only used if the option --batch has also
+        # been given. Since Version 2.1 the --pinentry-mode also needs to be set to loopback.
+        # see:https://www.gnupg.org/documentation/manuals/gnupg/GPG-Esoteric-Options.html#index-allow_002dsecret_002dkey_002dimport
+        gpg2 -v --pinentry-mode loopback --batch --secret-keyring ~/.gnupg/secring.gpg --import /tmp/private.key
 
-    logDebug "----- Imported list of GPG secret keys -----"
-    gpg2 --list-keys
-    gpg2 --list-secret-keys
+        logDebug "Cleanup to /tmp/private.key ..."
+        \rm -rf /tmp/private.key
+        ls -al ~/.gnupg/
+
+        logDebug "----- Imported list of GPG secret keys -----"
+        gpg2 --list-keys
+        gpg2 --list-secret-keys
+    fi
 
     # Notice: Test signing should be performed first to ensure that the gpg-agent service has been 
     # pre-started (gpg-agent --homedir /root/.gnupg --use-standard-socket --daemon), otherwise
     # an error may be reported : 'gpg: signing failed: Inappropriate ioctl for device'
-    logDebug "Preparing testing the GPG signing ..."
+    if [[ -z "$MAVEN_GPG_PASSPHRASE" ]]; then
+        logErr "The environment variable MAVEN_GPG_PASSPHRASE is missing."; exit 1
+    fi
+    logDebug "Prepare verifying the GPG signing ..."
     echo "test" | gpg2 -v --pinentry-mode loopback --passphrase $MAVEN_GPG_PASSPHRASE --clear-sign
 }
 
@@ -159,7 +164,7 @@ function do_build_deploy() {
     # or using:https://github.com/actions/setup-java/tree/v1.4.3#publishing-using-apache-maven
     do_configure_gpg
 
-    do_build_maven "-T 4C deploy -Prelease"
+    do_build_maven "--settings $BASE_DIR/.github/mvn-settings.xml -T 4C deploy -Prelease"
 }
 
 function do_push_image() {
@@ -167,11 +172,10 @@ function do_push_image() {
     local image_tag="$2"
     local image_registry="$3"
     if [[ -z "$DOCKERHUB_USERNAME" || -z "$DOCKERHUB_TOKEN" ]]; then
-      logWarn "The environment variable DOCKERHUB_USERNAME or DOCKERHUB_TOKEN is missing."
-      exit 1
+        logWarn "The environment variable DOCKERHUB_USERNAME or DOCKERHUB_TOKEN is missing."; exit 1
     fi
     if [ -z "$image_registry" ]; then
-      image_registry="docker.io"
+        image_registry="docker.io"
     fi
     logDebug "Login to $image_registry ..."
     docker login -u $DOCKERHUB_USERNAME -p $DOCKERHUB_TOKEN $image_registry
@@ -184,6 +188,9 @@ function do_push_image() {
 case $1 in
   version)
     print_pom_version
+    ;;
+  gpg-verify)
+    do_configure_gpg
     ;;
   build-maven)
     do_build_maven "-T 4C clean install"
